Refactor ChooserRowView out into its own file
